# expense-management-api/storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SpendWise API",
    "version": "1.0.0",
    "description": "API for managing expenses, budgets, balances and group contexts",
    "contact": {
      "email": "support@spendwise.com"
    }
  },
  "servers": [
    {
      "url": "http://expense-management-api.test//api",
      "description": "Local API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in format: Bearer <token>"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "List system and custom categories",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "context_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {
            "description": "Categories list"
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create custom category (Pro users only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["context_id", "name", "icon"],
                "properties": {
                  "context_id": {"type": "string", "format": "uuid"},
                  "name": {"type": "string"},
                  "icon": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created"
          }
        }
      }
    },
    "/categories/{category}": {
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete custom category",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "category", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {
            "description": "Category deleted"
          }
        }
      }
    },
    "/expenses": {
      "get": {
        "tags": ["Expenses"],
        "summary": "List expenses with filters",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "context_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"name": "date_from", "in": "query", "schema": {"type": "string", "format": "date"}},
          {"name": "date_to", "in": "query", "schema": {"type": "string", "format": "date"}},
          {"name": "category_id", "in": "query", "schema": {"type": "string", "format": "uuid"}},
          {"name": "keyword", "in": "query", "schema": {"type": "string"}},
          {"name": "per_page", "in": "query", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {
            "description": "Expenses list"
          }
        }
      },
      "post": {
        "tags": ["Expenses"],
        "summary": "Create new expense",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["context_id", "amount", "category_id"],
                "properties": {
                  "context_id": {"type": "string", "format": "uuid"},
                  "amount": {"type": "number", "example": 500},
                  "category_id": {"type": "string", "format": "uuid"},
                  "note": {"type": "string"},
                  "expense_date": {"type": "string", "format": "date"},
                  "split_type": {"type": "string", "enum": ["equal", "custom", "percentage"]},
                  "splits": {"type": "array", "items": {"type": "object"}}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense created"
          }
        }
      }
    },
    "/expenses/{expense}": {
      "get": {
        "tags": ["Expenses"],
        "summary": "Get expense details",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "expense", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {
            "description": "Expense details"
          }
        }
      },
      "put": {
        "tags": ["Expenses"],
        "summary": "Update expense",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "expense", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {"type": "number"},
                  "category_id": {"type": "string", "format": "uuid"},
                  "note": {"type": "string"},
                  "split_type": {"type": "string"},
                  "splits": {"type": "array"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense updated"
          }
        }
      },
      "delete": {
        "tags": ["Expenses"],
        "summary": "Delete expense",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "expense", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {
            "description": "Expense deleted"
          }
        }
      }
    },
    "/expenses/{expense}/settle": {
      "patch": {
        "tags": ["Expenses"],
        "summary": "Mark expense as settled",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "expense", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {
            "description": "Expense marked as settled"
          }
        }
      }
    },
...
  }
}